refactor: shared 컴포넌트 분리 및 적용
