+++ src/shared/ui/card/Card.tsx
@@ -1,0 +1,19 @@
+import * as React from "react";
+import { forwardRef } from "react";
+
+interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+}
+
+// 카드 컴포넌트
+// card UI
+export const Card = forwardRef<HTMLDivElement, CardProps>(
+  ({ className, ...props }, ref) => (
+    <div
+      ref={ref}
+      className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`}
+      {...props}
+    />
+  )
+);
+Card.displayName = "Card";
