+++ src/shared/ui/card/index.ts
@@ -1,0 +1,4 @@
+export { Card } from "./Card";
+export { CardHeader } from "./CardHeader";
+export { CardTitle } from "./CardTitle";
+export { CardContent } from "./CardContent";
