+++ src/shared/ui/dialog/DialogHeader.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+
+interface DialogHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+}
+
+export const DialogHeader: React.FC<DialogHeaderProps> = ({
+  className,
+  ...props
+}) => (
+  <div
+    className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`}
+    {...props}
+  />
+);
+DialogHeader.displayName = "DialogHeader";
