+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,4 @@
+export { Dialog, DialogOverlay, DialogPortal, DialogTrigger } from "./Dialog";
+export { DialogContent } from "./DialogContent";
+export { DialogHeader } from "./DialogHeader";
+export { DialogTitle } from "./DialogTitle";
