+++ src/shared/ui/index.tsx
@@ -1,402 +1,29 @@
-import * as React from "react";
-import { ComponentPropsWithoutRef, forwardRef } from "react";
-import * as SelectPrimitive from "@radix-ui/react-select";
-import * as DialogPrimitive from "@radix-ui/react-dialog";
-import { Check, ChevronDown, X } from "lucide-react";
-import { cva, VariantProps } from "class-variance-authority";
-
-/* ------------------------------------------------------------------
-  1) Button
------------------------------------------------------------------- */
-const buttonVariants = cva(
-  "inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background",
-  {
-    variants: {
-      variant: {
-        default: "bg-blue-500 text-white hover:bg-blue-600",
-        destructive: "bg-red-500 text-white hover:bg-red-600",
-        outline:
-          "border border-gray-300 bg-transparent text-gray-700 hover:bg-gray-100",
-        secondary: "bg-gray-200 text-gray-800 hover:bg-gray-300",
-        ghost: "bg-transparent text-gray-700 hover:bg-gray-100",
-        link: "underline-offset-4 hover:underline text-blue-500",
-      },
-      size: {
-        default: "h-10 py-2 px-4",
-        sm: "h-8 px-3 rounded-md text-xs",
-        lg: "h-11 px-8 rounded-md",
-        icon: "h-9 w-9",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-      size: "default",
-    },
-  }
-);
-
-// Button Props
-interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  className?: string;
-}
-
-// Button UI
-export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant, size, ...props }, ref) => {
-    return (
-      <button
-        className={buttonVariants({ variant, size, className })}
-        ref={ref}
-        {...props}
-      />
-    );
-  }
-);
-
-Button.displayName = "Button";
-
-/* ------------------------------------------------------------------
-  2) Input
------------------------------------------------------------------- */
-interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  className?: string;
-  type: string;
-}
-
-// Input UI
-export const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ className, type, ...props }, ref) => (
-    <input
-      type={type}
-      className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-      ref={ref}
-      {...props}
-    />
-  )
-);
-Input.displayName = "Input";
-
-/* ------------------------------------------------------------------
-  3) Card
------------------------------------------------------------------- */
-interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-// 카드 컴포넌트
-// Card UI
-export const Card = forwardRef<HTMLDivElement, CardProps>(
-  ({ className, ...props }, ref) => (
-    <div
-      ref={ref}
-      className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`}
-      {...props}
-    />
-  )
-);
-Card.displayName = "Card";
-
-interface CardHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-export const CardHeader = forwardRef<HTMLDivElement, CardHeaderProps>(
-  ({ className, ...props }, ref) => (
-    <div
-      ref={ref}
-      className={`flex flex-col space-y-1.5 p-6 ${className}`}
-      {...props}
-    />
-  )
-);
-CardHeader.displayName = "CardHeader";
-
-interface CardTitleProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-export const CardTitle = forwardRef<HTMLHeadingElement, CardTitleProps>(
-  ({ className, ...props }, ref) => (
-    <h3
-      ref={ref}
-      className={`text-2xl font-semibold leading-none tracking-tight ${className}`}
-      {...props}
-    />
-  )
-);
-CardTitle.displayName = "CardTitle";
-
-interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-export const CardContent = forwardRef<HTMLDivElement, CardContentProps>(
-  ({ className, ...props }, ref) => (
-    <div
-      ref={ref}
-      className={`p-6 pt-0 ${className}`}
-      {...props}
-    />
-  )
-);
-CardContent.displayName = "CardContent";
-
-/* ------------------------------------------------------------------
-  4) Textarea
------------------------------------------------------------------- */
-interface TextareaProps
-  extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
-  className?: string;
-}
-
-// 텍스트 영역 컴포넌트
-export const Textarea = forwardRef<HTMLTextAreaElement, TextareaProps>(
-  ({ className, ...props }, ref) => (
-    <textarea
-      className={`flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-      ref={ref}
-      {...props}
-    />
-  )
-);
-Textarea.displayName = "Textarea";
-
-/* ------------------------------------------------------------------
-  5) Select
------------------------------------------------------------------- */
-// 선택 컴포넌트
-export const Select = SelectPrimitive.Root;
-export const SelectGroup = SelectPrimitive.Group;
-export const SelectValue = SelectPrimitive.Value;
-
-// SelectTrigger
-type SelectTriggerProps = ComponentPropsWithoutRef<
-  typeof SelectPrimitive.Trigger
-> & {
-  className?: string;
-};
-
-export const SelectTrigger = forwardRef<HTMLButtonElement, SelectTriggerProps>(
-  ({ className, children, ...props }, ref) => (
-    <SelectPrimitive.Trigger
-      ref={ref}
-      className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-      {...props}
-    >
-      {children}
-      <ChevronDown className="h-4 w-4 opacity-50" />
-    </SelectPrimitive.Trigger>
-  )
-);
-SelectTrigger.displayName = SelectPrimitive.Trigger.displayName;
-
-type SelectContentProps = ComponentPropsWithoutRef<
-  typeof SelectPrimitive.Content
-> & {
-  className?: string;
-  position?: "itemAligned" | "popper";
-};
-
-export const SelectContent = forwardRef<HTMLDivElement, SelectContentProps>(
-  ({ className, children, position = "popper", ...props }, ref) => (
-    <SelectPrimitive.Portal>
-      <SelectPrimitive.Content
-        ref={ref}
-        className={`relative z-50 min-w-[8rem] overflow-hidden rounded-md border bg-white text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 ${className}`}
-        position={position}
-        {...props}
-      >
-        <SelectPrimitive.Viewport className="p-1">
-          {children}
-        </SelectPrimitive.Viewport>
-      </SelectPrimitive.Content>
-    </SelectPrimitive.Portal>
-  )
-);
-SelectContent.displayName = SelectPrimitive.Content.displayName;
-
-type SelectItemProps = ComponentPropsWithoutRef<typeof SelectPrimitive.Item> & {
-  className?: string;
-};
-
-export const SelectItem = forwardRef<HTMLDivElement, SelectItemProps>(
-  ({ className, children, ...props }, ref) => (
-    <SelectPrimitive.Item
-      ref={ref}
-      className={`relative flex w-full cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none focus:bg-accent focus:text-accent-foreground data-[disabled]:pointer-events-none data-[disabled]:opacity-50 ${className}`}
-      {...props}
-    >
-      <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
-        <SelectPrimitive.ItemIndicator>
-          <Check className="h-4 w-4" />
-        </SelectPrimitive.ItemIndicator>
-      </span>
-      <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
-    </SelectPrimitive.Item>
-  )
-);
-SelectItem.displayName = SelectPrimitive.Item.displayName;
-
-/* ------------------------------------------------------------------
-  6) Dialog
------------------------------------------------------------------- */
-// 대화상자 컴포넌트
-export const Dialog = DialogPrimitive.Root;
-export const DialogTrigger = DialogPrimitive.Trigger;
-export const DialogPortal = DialogPrimitive.Portal;
-export const DialogOverlay = DialogPrimitive.Overlay;
-
-type DialogContentProps = ComponentPropsWithoutRef<
-  typeof DialogPrimitive.Content
-> & {
-  className?: string;
-};
-
-export const DialogContent = forwardRef<HTMLDivElement, DialogContentProps>(
-  ({ className, children, ...props }, ref) => (
-    <DialogPortal>
-      <DialogOverlay className="fixed inset-0 z-50 bg-black/50 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0" />
-      <DialogPrimitive.Content
-        ref={ref}
-        className={`fixed left-[50%] top-[50%] z-50 grid w-full max-w-lg translate-x-[-50%] translate-y-[-50%] gap-4 border bg-white p-6 shadow-lg duration-200 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[48%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[48%] sm:rounded-lg md:w-full ${className}`}
-        {...props}
-      >
-        {children}
-        <DialogPrimitive.Close className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground">
-          <X className="h-4 w-4" />
-          <span className="sr-only">닫기</span>
-        </DialogPrimitive.Close>
-      </DialogPrimitive.Content>
-    </DialogPortal>
-  )
-);
-DialogContent.displayName = DialogPrimitive.Content.displayName;
-
-interface DialogHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-export const DialogHeader: React.FC<DialogHeaderProps> = ({
-  className,
-  ...props
-}) => (
-  <div
-    className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`}
-    {...props}
-  />
-);
-DialogHeader.displayName = "DialogHeader";
-
-type DialogTitleProps = ComponentPropsWithoutRef<
-  typeof DialogPrimitive.Title
-> & {
-  className?: string;
-};
-
-export const DialogTitle = forwardRef<HTMLHeadingElement, DialogTitleProps>(
-  ({ className, ...props }, ref) => (
-    <DialogPrimitive.Title
-      ref={ref}
-      className={`text-lg font-semibold leading-none tracking-tight ${className}`}
-      {...props}
-    />
-  )
-);
-DialogTitle.displayName = DialogPrimitive.Title.displayName;
-
-/* ------------------------------------------------------------------
-  7) Table
------------------------------------------------------------------- */
-interface TableProps extends React.TableHTMLAttributes<HTMLTableElement> {
-  className?: string;
-}
-// 테이블 컴포넌트
-export const Table = forwardRef<HTMLTableElement, TableProps>(
-  ({ className, ...props }, ref) => (
-    <div className="w-full overflow-auto">
-      <table
-        ref={ref}
-        className={`table-fixed w-full caption-bottom text-sm ${className}`}
-        {...props}
-      />
-    </div>
-  )
-);
-Table.displayName = "Table";
-
-interface TableHeaderProps
-  extends React.HTMLAttributes<HTMLTableSectionElement> {
-  className?: string;
-}
-
-export const TableHeader = forwardRef<
-  HTMLTableSectionElement,
-  TableHeaderProps
->(({ className, ...props }, ref) => (
-  <thead
-    ref={ref}
-    className={`[&_tr]:border-b ${className}`}
-    {...props}
-  />
-));
-TableHeader.displayName = "TableHeader";
-
-interface TableBodyProps extends React.HTMLAttributes<HTMLTableSectionElement> {
-  className?: string;
-}
-
-export const TableBody = forwardRef<HTMLTableSectionElement, TableBodyProps>(
-  ({ className, ...props }, ref) => (
-    <tbody
-      ref={ref}
-      className={`[&_tr:last-child]:border-0 ${className}`}
-      {...props}
-    />
-  )
-);
-TableBody.displayName = "TableBody";
-
-interface TableRowProps extends React.HTMLAttributes<HTMLTableRowElement> {
-  className?: string;
-}
-
-export const TableRow = forwardRef<HTMLTableRowElement, TableRowProps>(
-  ({ className, ...props }, ref) => (
-    <tr
-      ref={ref}
-      className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
-      {...props}
-    />
-  )
-);
-TableRow.displayName = "TableRow";
-
-interface TableHeadProps extends React.HTMLAttributes<HTMLTableCellElement> {
-  className?: string;
-}
-
-export const TableHead = forwardRef<HTMLTableCellElement, TableHeadProps>(
-  ({ className, ...props }, ref) => (
-    <th
-      ref={ref}
-      className={`h-10 px-2 text-left align-middle font-medium text-muted-foreground [&:has([role=checkbox])]:pr-0 ${className}`}
-      {...props}
-    />
-  )
-);
-TableHead.displayName = "TableHead";
-
-interface TableCellProps extends React.HTMLAttributes<HTMLTableCellElement> {
-  className?: string;
-}
-
-export const TableCell = forwardRef<HTMLTableCellElement, TableCellProps>(
-  ({ className, ...props }, ref) => (
-    <td
-      ref={ref}
-      className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`}
-      {...props}
-    />
-  )
-);
-TableCell.displayName = "TableCell";
+export { Button } from "./button";
+export { Input } from "./input";
+export { Card, CardHeader, CardTitle, CardContent } from "./card";
+export { Textarea } from "./textarea";
+export {
+  Select,
+  SelectGroup,
+  SelectValue,
+  SelectTrigger,
+  SelectContent,
+  SelectItem,
+} from "./select";
+export {
+  Dialog,
+  DialogOverlay,
+  DialogPortal,
+  DialogTrigger,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from "./dialog";
+export {
+  Table,
+  TableHeader,
+  TableBody,
+  TableRow,
+  TableHead,
+  TableCell,
+} from "./table";
