+++ src/shared/ui/select/index.ts
@@ -1,0 +1,4 @@
+export { Select, SelectGroup, SelectValue } from "./Select";
+export { SelectTrigger } from "./SelectTrigger";
+export { SelectContent } from "./SelectContent";
+export { SelectItem } from "./SelectItem";
