+++ src/shared/ui/table/TableHeader.tsx
@@ -1,0 +1,19 @@
+import * as React from "react";
+import { forwardRef } from "react";
+
+interface TableHeaderProps
+  extends React.HTMLAttributes<HTMLTableSectionElement> {
+  className?: string;
+}
+
+export const TableHeader = forwardRef<
+  HTMLTableSectionElement,
+  TableHeaderProps
+>(({ className, ...props }, ref) => (
+  <thead
+    ref={ref}
+    className={`[&_tr]:border-b ${className}`}
+    {...props}
+  />
+));
+TableHeader.displayName = "TableHeader";
